Grammar and Corrected Home Page
Fixed a minor grammar issue in the Algebra II page

Added the correct home page from local files. The only change is that the old version does not utilize BootStrap as effectively.
The home page should now shove the sidebar to the bottom if the window size is small enough (more optimal for mobile devices)